Linux: Detect when and old CEF version is installed and a new version is desired #1
CEF_VERSION now set the real CEF version holding + char.
CEF_VERSION is converted to URL percent encoding before download.
Read the README file and grep for the version if not found remove the
older folder and download the new CEF version.
This code has to be ported for Windows.